make improvents to ops.Serve
The sync.WaitGroup (which was ineffectual, see #77) has been replaced with an
error channel in preparation for #75 so that http.ListenAndServe's error can
eventually be thrown at the script that called Elsa.serve.
Renamed some variables to be more clear, also removed redundant comments.
Made the requests and responses channels unbuffered.